Move HorizontalLine to separate file
